refactor(solver): import RequestType from core.models, not bunking.models
Use the canonical pipeline-layer enum consistently in the solver layer,
matching how the pipeline layer's own code sources RequestType.
