+++ bunking/solver/direct_solver.py
@@ -13,13 +13,13 @@ from ortools.sat.python import cp_model
 
 from bunking.config import ConfigLoader
 from bunking.logging_config import get_logger
-from bunking.models import RequestType
 from bunking.models_v2 import (
     DirectBunkAssignment,
     DirectBunkRequest,
     DirectSolverInput,
     DirectSolverOutput,
 )
+from bunking.sync.bunk_request_processor.core.models import RequestType
 from bunking.sync.bunk_request_processor.shared.constants import SOURCE_FIELD_TO_CONFIG_KEY
 from campminder.client import get_current_season
