+++ bunking/solver/objective_evaluator.py
@@ -23,8 +23,8 @@ from typing import Any
 
 from bunking.config import ConfigLoader
 from bunking.logging_config import get_logger
-from bunking.models import RequestType
 from bunking.solver.bunk_ordering import get_bunk_rank
+from bunking.sync.bunk_request_processor.core.models import RequestType
 from bunking.sync.bunk_request_processor.shared.constants import SourceField
 
 logger = get_logger(__name__)
